# summarizer/docs/api-spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Summarizer API",
        "version": "1.0.0"
    },
    "paths": {
        "/summarize": {
            "post": {
                "summary": "文字起こしを要約ジョブとして登録",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "text": {
                                        "type": "string"
                                    },
                                    "strength": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "maximum": 5,
                                        "example": 3
                                    }
                                },
                                "required": [
                                    "text",
                                    "strength"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "要約ジョブ登録成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "request_id": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "リクエスト不正",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "サーバー内部エラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/summarize/{request_id}": {
            "get": {
                "summary": "要約結果またはステータスを取得",
                "parameters": [
                    {
                        "in": "path",
                        "name": "request_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "処理中または完了",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/StatusResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ResultResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "存在しないまたは期限切れ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "不正なリクエスト",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "サーバー内部エラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StatusResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "working",
                            "error"
                        ]
                    },
                    "error": {
                        "type": "string",
                        "nullable": true
                    }
                }
            },
            "ResultResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "done"
                        ]
                    },
                    "summary": {
                        "type": "string"
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            }
        }
    }
}